third_party/blink/web_tests/external/wpt/html/browsers/browsing-the-web/navigating-across-documents/replace-before-load/form-submit-popup-expected.txt

This is a testharness.js-based test.
FAIL No replace before load, triggered by formElement.submit() in the opener window, after the opener has loaded assert_equals: window's after-submit history.length expected 2 but got 1
Harness: the test ran to completion.
